requirements.txt
Django==2.2.2
djangorestframework==3.9.4
numpy==1.16.4
requests==2.22.0
vector_api_folder/
